{"schema":"decodex/commit/1","summary":"split json rpc tests into modules","authority":"manual"}
