+++ apps/decodex/src/agent/json_rpc/tests.rs
@@ -1,0 +1,4 @@
+mod environment;
+mod support;
+mod transport;
+mod wire;
